Renaming pandas to pandas_magic to avoid circular import problem (#68)
